+++ R/class.R
@@ -116,6 +116,7 @@ setClass("tinytable_tabularray", contains = "tinytable")
 setClass("tinytable_bootstrap", contains = "tinytable")
 setClass("tinytable_typst", contains = "tinytable")
 setClass("tinytable_grid", contains = "tinytable")
+setClass("tinytable_dataframe", contains = "tinytable")
 
 #' Apply style settings to a tinytable
 #'