+++ R/style_grid_dataframe.R
@@ -1,0 +1,25 @@
+#' tinytable S4 method
+#' 
+#' @keywords internal
+setMethod(f = "style_eval",
+          signature = "tinytable_dataframe",
+          definition = style_grid_internal
+)
+
+
+#' tinytable S4 method
+#' 
+#' @keywords internal
+setMethod(f = "group_eval",
+          signature = "tinytable_dataframe",
+          definition = identity
+)
+
+
+#' tinytable S4 method
+#' 
+#' @keywords internal
+setMethod(f = "finalize",
+          signature = "tinytable_dataframe",
+          definition = identity
+)